requirements.txt
# Core Python dependencies
python-dotenv==1.0.0
streamlit
requests
pandas

# Google Generative AI (Gemini)
google-genai

# Google Sheets integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib
google-auth-httplib2
